Refactor and tests
