+++ apptax/tests/assets/import_media.csv
@@ -2,4 +2,5 @@ cd_ref,Nom Complet,Nom Valide,Nom Vern
 60457,"Nyctalus lasiopterus (Schreber, 1780)","Nyctalus lasiopterus (Schreber, 1780)",Grande Noctule
 79300,"Myotis brandtii (Eversmann, 1845)","Myotis brandtii (Eversmann, 1845)",Murin de Brandt
 60557,"Tadarida teniotis (Rafinesque, 1814)","Tadarida teniotis (Rafinesque, 1814)",Molosse de Cestoni
+4117,"Turdus merula Linnaeus, 1758","Turdus merula Linnaeus, 1758",Merle noir
 invalid,invalid name,,
